Most of the rules done. TODO: entry point, output, string formatting

+++ Backend/drools-spring-app/src/main/java/sbnz/integracija/example/facts/Book.java
@@ -22,7 +22,7 @@ public class Book implements Serializable{
 	@Column
 	private int avaivableNo;
 	
-	private int searchMatch=0;
+	private double searchMatch=0;
 	
 	public Long getId() {
 		return id;
@@ -60,10 +60,10 @@ public class Book implements Serializable{
 	public void setAvaivableNo(int avaivableNo) {
 		this.avaivableNo = avaivableNo;
 	}
-	public int getMatch() {
+	public double getMatch() {
 		return searchMatch;
 	}
-	public void setMatch(int match) {
+	public void setMatch(double match) {
 		this.searchMatch = match;
 	}
